+++ src/app/core/guards/auth/auth.guard.ts
@@ -1,12 +1,11 @@
-import { CanActivateFn, Router } from '@angular/router';
-import { inject } from '@angular/core';
-import { AuthenticationService } from '../../../auth/authentication.service';
-import { map } from 'rxjs';
-import { CodeStatus } from '../../enums/CodeStatus';
-import { MessageDTO } from '../../interfaces/MessageDTO';
-import { catchError } from 'rxjs/operators';
-import { of } from 'rxjs';
-import { HttpResponse } from '@angular/common/http';
+import {CanActivateFn, Router} from '@angular/router';
+import {inject} from '@angular/core';
+import {AuthenticationService} from '../../../auth/authentication.service';
+import {map, of} from 'rxjs';
+import {CodeStatus} from '../../types/enums/CodeStatus';
+import {MessageDTO} from '../../types/other/MessageDTO';
+import {catchError} from 'rxjs/operators';
+import {HttpResponse} from '@angular/common/http';
 
 export const authGuard: CanActivateFn = (route, state) => {
   const authService = inject(AuthenticationService);
@@ -14,7 +13,7 @@ export const authGuard: CanActivateFn = (route, state) => {
 
   return authService.isAuthenticated().pipe(
     map((response: HttpResponse<MessageDTO>) => {
-      const message = response.body ?? { status: CodeStatus.OK, message: '' };
+      const message = response.body ?? {status: CodeStatus.OK, message: ''};
       if (message.status === CodeStatus.OK) {
         return true;
       }
