+++ src/app/core/services/attendance/attendance.service.spec.ts
@@ -1,6 +1,6 @@
-import { TestBed } from '@angular/core/testing';
+import {TestBed} from '@angular/core/testing';
 
-import { AttendanceService } from './attendance.service';
+import {AttendanceService} from './attendance.service';
 
 describe('AttendanceService', () => {
   let service: AttendanceService;
