+++ src/app/core/services/breadcrumbs/breadcrumb.service.spec.ts
@@ -1,6 +1,6 @@
-import { TestBed } from '@angular/core/testing';
+import {TestBed} from '@angular/core/testing';
 
-import { BreadcrumbService } from './breadcrumb.service';
+import {BreadcrumbService} from './breadcrumb.service';
 
 describe('BreadcrumbService', () => {
   let service: BreadcrumbService;
