+++ src/app/core/services/student/student.service.spec.ts
@@ -1,6 +1,6 @@
-import { TestBed } from '@angular/core/testing';
+import {TestBed} from '@angular/core/testing';
 
-import { StudentService } from './student.service';
+import {StudentService} from './student.service';
 
 describe('StudentService', () => {
   let service: StudentService;
